master: converted return type of processor utilization
* Converted the return type of processor utilization from double to
float.

+++ include/linux_parser.h
@@ -47,7 +47,7 @@ namespace LinuxParser {
         kGuestNice_
     };
 
-    std::vector<double> CpuUtilization();
+    std::vector<float> CpuUtilization();
 
     long Jiffies();
